Modify job driver/work giver re: new designations
Both the job driver and work giver classes now respect the new
designations for combat dummies. The mod will use the type of weapon
currently being held by the pawn when trying to find suitable dummies -
swapping to a different type of weapon after a dummy has been found is
only allowed on dummies with the legacy functionality (for backwards
compatibility).

Works on Kriil/rimworld-combat-training#2.