Fix checks for designations
To prevent a designation check from preventing a pawn to reclaim their
original weapon, the designation checks have been moved from a FailOn
call to a check at the start of the job, and one in the doneTraining
check. This runs less-often than FailOn, but appears to have no
hugely-detrimental impact.

Works on Kriil/rimworld-combat-training#2.